feat: runing multi mode depends on argvs
