+++ package.json
@@ -6,9 +6,9 @@
   "type": "module",
   "scripts": {
     "test": "jest",
-    "start": "npx ts-node ./src/index.ts",
+    "start": "npx ts-node src/index.ts",
     "start:dev": "npx nodemon",
-    "start:multi": "npx ts-node ./src/multi.ts",
+    "start:multi": "npx ts-node src/index.ts --multi",
     "start:prod": "rimraf ./build && npx tsc"
   },
   "author": "emp74ark",
