+++ src/controller/data.controller.ts
@@ -1,5 +1,6 @@
 import { User } from '../interfaces.js';
-import { sharedData } from '../model/user.model.js';
+import { sharedData, updateSharedData } from '../model/user.model.js';
+import cluster from 'cluster';
 
 export function getData() {
   process.send?.({ type: 'get' }); // from worker to master
@@ -8,4 +9,7 @@ export function getData() {
 
 export function updateData(values: User[]) {
   process.send?.({ type: 'update', payload: values }); // from worker to master
+  if (cluster.isPrimary){
+    updateSharedData(values)
+  }
 }
