+++ src/modules/helpers/multiChecker.ts
@@ -1,0 +1,4 @@
+export function multiModeChecker() {
+  const args = process.argv;
+  return args.slice(2)[0] === '--multi';
+}
